edit: navigator add
